Register Copilot CLI debug log/state file commands in workbench
Move 'Export Agent Host Debug Logs' and 'Open Copilot CLI State File'
into vs/workbench/contrib/chat so they're available in vscode as well as
the agents app.

The shared collection/resolution logic lives in workbench helpers; the
two layers each register their own Action2 with appropriate session
discovery (IChatWidgetService in workbench, ISessionsManagementService
in sessions) and gating.
