+++ src/vs/workbench/contrib/chat/browser/actions/chatDeveloperActions.ts
@@ -20,6 +20,7 @@ import { IChatWidgetService } from '../chat.js';
 import { IStorageService, StorageScope } from '../../../../../platform/storage/common/storage.js';
 import { AUTOPILOT_DONT_SHOW_AGAIN_KEY, AUTO_APPROVE_DONT_SHOW_AGAIN_KEY } from '../../common/chatPermissionStorageKeys.js';
 import { resetShownWarnings } from '../../common/chatPermissionWarnings.js';
+import { OpenCopilotCliStateFileAction } from './openCopilotCliStateFileAction.js';
 
 function uriReplacer(_key: string, value: unknown): unknown {
 	if (URI.isUri(value)) {
@@ -41,6 +42,7 @@ export function registerChatDeveloperActions() {
 	registerAction2(InspectChatModelReferencesAction);
 	registerAction2(ClearRecentlyUsedLanguageModelsAction);
 	registerAction2(ResetChatPermissionWarningDialogsAction);
+	registerAction2(OpenCopilotCliStateFileAction);
 }
 
 function formatChatModelReferenceInspection(accessor: ServicesAccessor): string {
