+++ src/vs/workbench/contrib/chat/electron-browser/actions/chatDeveloperActions.ts
@@ -11,10 +11,12 @@ import { INativeHostService } from '../../../../../platform/native/common/native
 import { ChatContextKeys } from '../../common/actions/chatContextKeys.js';
 import { IChatService } from '../../common/chatService/chatService.js';
 import { DebugAgentHostInDevToolsAction } from './debugAgentHostAction.js';
+import { ExportAgentHostDebugLogsAction } from './exportAgentHostDebugLogsAction.js';
 
 export function registerChatDeveloperActions() {
 	registerAction2(OpenChatStorageFolderAction);
 	registerAction2(DebugAgentHostInDevToolsAction);
+	registerAction2(ExportAgentHostDebugLogsAction);
 }
 
 class OpenChatStorageFolderAction extends Action2 {
